Restructure angularjs-portal-home into require packages and modules, following how angular modules are defined
