+++ angularjs-portal-home/src/main/webapp/home.html
@@ -52,34 +52,7 @@
 
   <!-- FOOTER  -->
   <site-footer></site-footer>
-  <script type="text/javascript" src="js/config.js"></script>
-  <script type="text/javascript" src="bower_components/jquery/dist/jquery.min.js"></script>
-  <script type="text/javascript" src="bower_components/jquery-ui/jquery-ui.min.js"></script>
-  <script type="text/javascript" src="bower_components/angular/angular.min.js"></script>
-  <script type="text/javascript" src="bower_components/angular-route/angular-route.min.js"></script>
-  <script type="text/javascript" src="bower_components/angular-sanitize/angular-sanitize.min.js"></script>
-  <script type="text/javascript" src="bower_components/ngstorage/ngStorage.min.js"></script>
-  <script type="text/javascript" src="js/sortable.js"></script>
-  <script type="text/javascript" src="js/portal.js"></script>
-  <script type="text/javascript" src="js/main/main-controllers.js"></script>
-  <script type="text/javascript" src="js/main/search-controller.js"></script>
-  <script type="text/javascript" src="js/main/main-service.js"></script>
-  <script type="text/javascript" src="js/main/main-directives.js"></script>
-  <script type="text/javascript" src="js/layout/widget-controllers.js"></script>
-  <script type="text/javascript" src="js/layout/layout-controller.js"></script>
-  <script type="text/javascript" src="js/layout/layout-services.js"></script>
-  <script type="text/javascript" src="js/layout/layout-directives.js"></script>
-  <script type="text/javascript" src="js/portal/misc/controllers.js"></script>
-  <script type="text/javascript" src="js/portal/misc/services.js"></script>
-  <script type="text/javascript" src="js/notifications/notification-controllers.js"></script>
-  <script type="text/javascript" src="js/notifications/notification-directives.js"></script>
-  <script type="text/javascript" src="js/app-config.js"></script>
-  <script type="text/javascript" src="js/portal/misc/directives.js"></script>
-  <script type="text/javascript" src="js/portal/misc/filters.js"></script>
-  <script type="text/javascript" src="js/marketplace/marketplace-service.js"></script>
-  <script type="text/javascript" src="js/marketplace/marketplace-controller.js"></script>
-  <script type="text/javascript" src="js/marketplace/marketplace-directive.js"></script>
-  <script type="text/javascript" src="js/ui-bootstrap.js"></script>
   <script type="text/javascript" src="js/ga.js"></script>
+  <script type="text/javascript" src="bower_components/requirejs/require.js" data-main="main"></script>
 </body>
 </html>
