+++ angularjs-portal-home/src/main/webapp/js/app-config.js
@@ -1,16 +1,19 @@
+define(['angular'], function(angular){
 
-(function() {
-angular.module('app-config', [])
-.constant('APP_FLAGS', {
-                          'enableToggle': true,
-                          'defaultView' : 'list',
-                          'list' : true,
-                          'widgets' : true
-                       }
-          )
-.constant('SERVICE_LOC', {
-                           'sessionInfo' : '/portal/web/session.json',
-                           'sidebarInfo' : '/web/samples/sidebar.json'
-                       });
+    var app = angular.module('app-config', []);
+    app
+        .constant('APP_FLAGS', {
+            'enableToggle': true,
+            'defaultView' : 'list',
+            'list' : true,
+            'widgets' : true
+        })
+        .constant('SERVICE_LOC', {
+            'sessionInfo' : '/portal/web/session.json',
+            'sidebarInfo' : '/web/samples/sidebar.json'
+        });
 
-})();
+    return app;
+
+});
+
