+++ angularjs-portal-home/src/main/webapp/portal/layout/partials/default-card.html
