+++ angularjs-portal-home/src/main/webapp/portal/layout/partials/example-widgets.html
