+++ angularjs-portal-home/src/main/webapp/portal/layout/partials/home-toggle.html
