+++ angularjs-portal-home/src/main/webapp/portal/layout/partials/marketplace-light.html
