+++ angularjs-portal-home/src/main/webapp/portal/layout/partials/portlet-icon.html
