+++ angularjs-portal-home/src/main/webapp/portal/layout/partials/static-content-max.html
