+++ angularjs-portal-home/src/main/webapp/portal/layout/partials/widget-card.html
