+++ angularjs-portal-home/src/main/webapp/portal/layout/widget/partials/option-link.html
