+++ angularjs-portal-home/src/main/webapp/portal/layout/widget/partials/weather.html
