+++ angularjs-portal-home/src/main/webapp/portal/marketplace/directives.js
@@ -1,0 +1,23 @@
+'use strict';
+
+define(['angular', 'require'], function(angular, require) {
+
+    var app = angular.module('portal.marketplace.directives', []);
+    app.directive('marketplacePortlet', function() {
+        return {
+            restrict : 'E',
+            templateUrl : require.toUrl('./partials/marketplace-portlet.html')
+        }
+    });
+
+    app.directive('ratingModalTemplate', function() {
+        return {
+            restrict : 'E',
+            templateUrl : require.toUrl('./partials/rating-modal-template.html')
+        }
+    });
+
+    return app;
+
+});
+
