+++ angularjs-portal-home/src/main/webapp/portal/marketplace/partials/marketplace-portlet.html
