+++ angularjs-portal-home/src/main/webapp/portal/marketplace/partials/rating-modal-template.html
