+++ angularjs-portal-home/src/main/webapp/portal/search/controllers.js
@@ -1,0 +1,46 @@
+'use strict';
+
+define(['angular'], function(angular) {
+
+    var app = angular.module('portal.search.controllers', []);
+    app.controller('SearchController', [ 'marketplaceService', '$location', '$scope', '$localStorage', function(marketplaceService, $location, $scope, $localStorage) {
+        $scope.initialFilter = '';
+        $scope.filterMatches = [];
+        $scope.portletListLoading = true;
+        if($localStorage && $localStorage.typeaheadSearch) {
+            marketplaceService.getPortlets().then(function(data){
+                $scope.portlets = data.portlets;
+                $scope.portletListLoading = false;
+            });
+        }
+
+        $scope.$watch('initialFilter', function(newVal, oldVal) {
+            if (!newVal || !$scope.portlets) {
+                $scope.filterMatches = [];
+                return;
+            }
+
+            $scope.filterMatches = marketplaceService.filterPortletsBySearchTerm($scope.portlets, newVal);
+        });
+
+        $scope.onSelect = function(portlet) {
+            $location.path("/apps/details/"+ portlet.fname);
+            $scope.initialFilter = "";
+            $scope.showSearch = false;
+            $scope.showSearchFocus = false;
+        };
+
+        $scope.submit = function(){
+            if($scope.initialFilter != "") {
+                $location.path("/apps/search/"+ $scope.initialFilter);
+                $scope.initialFilter = "";
+                $scope.showSearch = false;
+                $scope.showSearchFocus = false;
+            }
+        };
+    }]);
+
+    return app;
+
+});
+
